benchmark: extract logic to separate package
